+++ context-menu.js
@@ -20,15 +20,19 @@ export function showContextMenu(e, items) {
             ? el('i', { className: `ptmt-context-menu-icon ${item.icon}` })
             : el('span', { className: 'ptmt-context-menu-icon' }, item.icon || '');
 
-        const menuItem = el('div', { className: 'ptmt-context-menu-item' },
+        const menuItem = el('div', { className: `ptmt-context-menu-item${item.disabled ? ' ptmt-context-menu-item-disabled' : ''}` },
             iconElement,
             el('span', {}, item.label)
         );
 
-        menuItem.addEventListener('click', () => {
-            item.onClick();
-            hideContextMenu();
-        });
+        if (item.disabled) {
+            if (item.disabledTitle) menuItem.title = item.disabledTitle;
+        } else {
+            menuItem.addEventListener('click', () => {
+                item.onClick();
+                hideContextMenu();
+            });
+        }
 
         menu.appendChild(menuItem);
     });